backend/requirements.txt
fastapi
pydantic<2.0
uvicorn[standard]
fal-serverless
Pillow